Add TimeTravel to testing
  * Remove thread sleep
  * Add Date generator to mock new date times for delays
  * Add remove trigger test

+++ OneSignalSDK/onesignal/src/main/java/com/onesignal/OSInAppMessageRepository.java
@@ -82,7 +82,7 @@ class OSInAppMessageRepository {
                     String messageId = cursor.getString(cursor.getColumnIndex(OneSignalDbContract.InAppMessageTable.COLUMN_NAME_MESSAGE_ID));
                     String clickIds = cursor.getString(cursor.getColumnIndex(OneSignalDbContract.InAppMessageTable.COLUMN_CLICK_IDS));
                     int displayQuantity = cursor.getInt(cursor.getColumnIndex(OneSignalDbContract.InAppMessageTable.COLUMN_NAME_DISPLAY_QUANTITY));
-                    double lastDisplay = cursor.getDouble(cursor.getColumnIndex(OneSignalDbContract.InAppMessageTable.COLUMN_NAME_LAST_DISPLAY));
+                    long lastDisplay = cursor.getLong(cursor.getColumnIndex(OneSignalDbContract.InAppMessageTable.COLUMN_NAME_LAST_DISPLAY));
 
                     JSONArray clickIdsArray = new JSONArray(clickIds);
                     Set<String> clickIdsSet = new HashSet<>();
