+++ OneSignalSDK/unittest/src/test/java/com/onesignal/OneSignalPackagePrivateHelper.java
@@ -362,12 +362,12 @@ public class OneSignalPackagePrivateHelper {
       }
 
       @Override
-      public double getLastDisplayTime() {
+      public long getLastDisplayTime() {
          return this.displayStats.getLastDisplayTime();
       }
 
       @Override
-      public void setLastDisplayTime(double lastDisplayTime) {
+      public void setLastDisplayTime(long lastDisplayTime) {
          this.displayStats.setLastDisplayTime(lastDisplayTime);
       }
 
@@ -397,12 +397,12 @@ public class OneSignalPackagePrivateHelper {
       }
 
       @Override
-      public double getDisplayDelay() {
+      public long getDisplayDelay() {
          return this.displayStats.getDisplayDelay();
       }
 
       @Override
-      public void setDisplayDelay(double displayDelay) {
+      public void setDisplayDelay(long displayDelay) {
          this.displayStats.setDisplayDelay(displayDelay);
       }
